Remove rate policing from tc filter.
Currently ATC rate limits both in the htb qdisc and in the filter. I've seen unreliable results,
especially when tested with something like speedtest. Often the throughput remains too low, 
or it is very bursty. Removing rate-limiting from the filter and keeping it only in htb seems to make 
things better.